+++ src/types.d.ts
@@ -1,0 +1,9 @@
+type NotionData = Record<string, unknown>
+type NotionState = {
+  id: string
+  data: NotionData
+}
+type Message = {
+  command: string
+  text: string
+}